python_modules/requirements.scrapy.txt
# General Django installation.
Django==3.0.7
psycopg2-binary==2.8.6
channels==2.4.0
channels-redis==3.0.0

# Django extensions.
django-bootstrap4==2.1.1
django-crispy-forms==1.9.2
django-ipware==2.1.0
django-dynamic-preferences==1.9
better-profanity==0.6.1

# Testing.
pytest-django==3.10.0

# Redis.
redis==3.4.1
python-redis-lock==3.6.0

# Scrapy.
Scrapy==2.2.1
simpleflock==0.0.3
# (for Selenium and limiting access to it):
Selenium==3.141.0
python-redis-lock==3.6.0
